+++ src/MessageStateEvents.ts
@@ -17,24 +17,15 @@ export interface IMessageStateEventsOptions {
 };
 
 export default class MessageStateEvents<T> implements IStateEvents<T> {
-  /** @deprecated For internal use only. */
-  instanceId: string;
-  /** @deprecated Use `getCurrent()` instead. */
-  current: T;
-  /** @deprecated For internal use only. */
-  name: string;
-  /** @deprecated For internal use only. */
-  initTimer: ReturnType<typeof setTimeout> | null;
-  /** @deprecated For internal use only. */
-  timestamp: number | null;
-  /** @deprecated For internal use only. */
-  callbacks: {callback: ICallback<T>, onError: IErrorCallback | null}[];
-  /** @deprecated For internal use only. */
-  handler: {callback: ICallback<T>, wrappedCallback: IMessageHandler, onError: IErrorCallback | null, debugHandler: DebugListener<IDebugEvent<T>> | null} | null;
-  /** @deprecated For internal use only. */
-  targets: Target[];
-  /** @deprecated For internal use only. */
-  allowDebug: boolean;
+  private instanceId: string;
+  private current: T;
+  private name: string;
+  private initTimer: ReturnType<typeof setTimeout> | null;
+  private timestamp: number | null;
+  private callbacks: {callback: ICallback<T>, onError: IErrorCallback | null}[];
+  private handler: {callback: ICallback<T>, wrappedCallback: IMessageHandler, onError: IErrorCallback | null, debugHandler: DebugListener<IDebugEvent<T>> | null} | null;
+  private targets: Target[];
+  private allowDebug: boolean;
   constructor(initial: T, name: string, options: IMessageStateEventsOptions | boolean = {}, allowDebug = false) {
     this.instanceId = btoa(String.fromCharCode(...crypto.getRandomValues(new Uint8Array(8))));
     this.current = initial;
@@ -231,11 +222,11 @@ export default class MessageStateEvents<T> implements IStateEvents<T> {
     }
   }
 
-  isInitialized() {
+  private isInitialized() {
     return this.timestamp !== null;
   }
 
-  initialize() {
+  private initialize() {
     if (this.initTimer !== null) {
       clearTimeout(this.initTimer);
       this.initTimer = null;
@@ -243,7 +234,7 @@ export default class MessageStateEvents<T> implements IStateEvents<T> {
     }
   }
 
-  uninitialize() {
+  private uninitialize() {
     if (this.initTimer !== null) {
       clearTimeout(this.initTimer);
       this.initTimer = null;