+++ src/StateEventsHook.ts
@@ -1,7 +1,6 @@
 import type { IStateEvents, IErrorCallback } from './types/StateEvents';
 
 import { useState, useEffect } from 'react';
-import * as PropTypes from 'prop-types';
 
 const useStateEvents = <T>(stateEvents: IStateEvents<T>, onError: IErrorCallback | null = null): [T, (state: T) => void] => {
   const [value, setValue] = useState(stateEvents.getCurrent());
@@ -15,10 +14,5 @@ const useStateEvents = <T>(stateEvents: IStateEvents<T>, onError: IErrorCallback
   const newSetValue = (state: T) => stateEvents.publish(state);
   return [value, newSetValue];
 }
-
-useStateEvents.propTypes = {
-  stateEvents: PropTypes.object.isRequired,
-  onError: PropTypes.func
-};
 
 export default useStateEvents;