+++ src/index.ts
@@ -1,10 +1,9 @@
 import { debugAnnounce, debugAddListener, debugRemoveListener, debugSend } from './debugHelper';
 import LocalStateEvents from './LocalStateEvents';
-import ExternalStateEvents from './ExternalStateEvents';
 import MessageStateEvents from './MessageStateEvents';
 import useStateEvents from './StateEventsHook';
 import Subscription from './Subsciption';
 
-export { debugAnnounce, debugAddListener, debugRemoveListener, debugSend, LocalStateEvents, ExternalStateEvents, MessageStateEvents, useStateEvents, Subscription };
+export { debugAnnounce, debugAddListener, debugRemoveListener, debugSend, LocalStateEvents, MessageStateEvents, useStateEvents, Subscription };
 export * from './types/StateEvents';
 export * from './types/EventHandlers';